all news to non logined user added

+++ src/main/java/com/chirag/news/controller/MyNewsHomeController.java
@@ -1,6 +1,7 @@
 package com.chirag.news.controller;
 
 import com.chirag.news.model.DTO.NewsDTO;
+import com.chirag.news.service.JwtTokenService;
 import com.chirag.news.service.MyHomeNewsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -11,7 +12,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.client.HttpClientErrorException;
 
 import java.util.List;
 
@@ -20,10 +20,16 @@ public class MyNewsHomeController {
 
     @Autowired
     private MyHomeNewsService myHomeNewsService;
+    @Autowired
+    private JwtTokenService jwtTokenService;
 
     @ResponseBody
     @GetMapping("/home/my-news")
-    public List<NewsDTO> getNews(@RequestHeader("username") String username) throws Exception {
+    public List<NewsDTO> getNews(@RequestHeader("username") String username,
+                                 @RequestHeader("x-jwt-token") String jwt) throws Exception {
+        if(jwtTokenService.verifyS2sJwtToken(jwt)!=true) {
+            throw new Exception("jwt-token not authenticated");
+        }
         return myHomeNewsService.getNews(username);
     }
 
@@ -31,33 +37,53 @@ public class MyNewsHomeController {
     @PostMapping("/home/my-news/edit/{id}")
     public ResponseEntity updateNews(@PathVariable("id") Long id,
                                      @RequestHeader("username") String username,
-                                     @RequestParam("news_body") String newsBody) throws Exception {
+                                     @RequestParam("news_body") String newsBody,
+                                     @RequestHeader("x-jwt-token") String jwt) throws Exception {
+        if(jwtTokenService.verifyS2sJwtToken(jwt)!=true) {
+            throw new Exception("jwt-token not authenticated");
+        }
         return myHomeNewsService.updateNews(id,newsBody,username);
     }
 
     @ResponseBody
     @PostMapping("/home/my-news/add")
-    public ResponseEntity addNews(@RequestHeader("username") String username,
-                                  @RequestParam("news_body") String newsBody) throws Exception {
+    public NewsDTO addNews(@RequestHeader("username") String username,
+                           @RequestParam("news_body") String newsBody,
+                           @RequestHeader("x-jwt-token") String jwt) throws Exception {
+        if(jwtTokenService.verifyS2sJwtToken(jwt)!=true) {
+            throw new Exception("jwt-token not authenticated");
+        }
         return myHomeNewsService.addNews(username, newsBody);
     }
 
     @ResponseBody
     @PostMapping("/home/my-news/delete/{id}")
     public Boolean deleteNews(@PathVariable("id") Long id,
-                              @RequestHeader("username") String username){
+                              @RequestHeader("username") String username,
+                              @RequestHeader("x-jwt-token") String jwt) throws Exception {
+        if(jwtTokenService.verifyS2sJwtToken(jwt)!=true) {
+            throw new Exception("jwt-token not authenticated");
+        }
         return myHomeNewsService.deleteNews(id,username);
     }
 
     @ResponseBody
     @GetMapping("/home/my-news/bookmarked")
-    public List<NewsDTO> bookmarkedNews(@RequestHeader("username")  String username) throws Exception {
+    public List<NewsDTO> bookmarkedNews(@RequestHeader("username")  String username,
+                                        @RequestHeader("x-jwt-token") String jwt) throws Exception {
+        if(jwtTokenService.verifyS2sJwtToken(jwt)!=true) {
+            throw new Exception("jwt-token not authenticated");
+        }
         return myHomeNewsService.bookmarkedNews(username);
     }
 
     @ResponseBody
     @GetMapping("/home/my-news/liked")
-    public List<NewsDTO> likedNews(@RequestHeader("username") String username) throws Exception {
+    public List<NewsDTO> likedNews(@RequestHeader("username") String username,
+                                   @RequestHeader("x-jwt-token") String jwt) throws Exception {
+        if(jwtTokenService.verifyS2sJwtToken(jwt)!=true) {
+            throw new Exception("jwt-token not authenticated");
+        }
         return myHomeNewsService.likedNews(username);
     }
 
@@ -65,7 +91,11 @@ public class MyNewsHomeController {
     @PostMapping("/home/my-news/bookmark-news/{id}")
     public Boolean bookmark(@RequestHeader("username") String username,
                             @PathVariable("id") Long id,
-                            @RequestParam(value = "bookmark", required = false, defaultValue = "1") Integer bookmark) throws Exception {
+                            @RequestParam(value = "bookmark", required = false, defaultValue = "1") Integer bookmark,
+                            @RequestHeader("x-jwt-token") String jwt) throws Exception {
+        if(jwtTokenService.verifyS2sJwtToken(jwt)!=true) {
+            throw new Exception("jwt-token not authenticated");
+        }
         if(bookmark==1 || bookmark==0){
           return myHomeNewsService.bookmarkNews(username, id, bookmark);
         }else{
@@ -76,8 +106,12 @@ public class MyNewsHomeController {
     @ResponseBody
     @PostMapping("/home/my-news/like-news/{id}")
     public Long like(@RequestHeader("username") String username,
-                        @PathVariable("id") Long id,
-                        @RequestParam(value = "like", required = false,defaultValue = "1") Integer like) throws Exception {
+                     @PathVariable("id") Long id,
+                     @RequestParam(value = "like", required = false,defaultValue = "1") Integer like,
+                     @RequestHeader("x-jwt-token") String jwt) throws Exception {
+        if(jwtTokenService.verifyS2sJwtToken(jwt)!=true) {
+            throw new Exception("jwt-token not authenticated");
+        }
         if(like==1 || like==0){
             return myHomeNewsService.likeNews(username, id, like);
         }else{
